+++ nodeconductor/core/managers.py
@@ -45,6 +45,9 @@ class GenericKeyMixin(object):
             kwargs[self.object_id_field] = generic_key_value.id
             generic_key_content_type = ContentType.objects.get_for_model(generic_key_value)
             kwargs[self.content_type_field] = generic_key_content_type
+        elif suffix == 'in':
+            kwargs[self.object_id_field + '__in'] = [obj.id for obj in generic_key_value]
+            kwargs[self.content_type_field] = ContentType.objects.get_for_model(generic_key_value[0])
         elif suffix == 'isnull':
             kwargs[self.object_id_field + '__isnull'] = generic_key_value
             kwargs[self.content_type_field + '__isnull'] = generic_key_value