+++ nodeconductor/cost_tracking/serializers.py
@@ -14,11 +14,12 @@ class PriceEstimateSerializer(AugmentedSerializerMixin, serializers.HyperlinkedM
     scope = GenericRelatedField(related_models=models.PriceEstimate.get_editable_estimated_models())
     scope_name = serializers.SerializerMethodField()
     scope_type = serializers.SerializerMethodField()
+    resource_type = serializers.SerializerMethodField()
 
     class Meta(object):
         model = models.PriceEstimate
-        fields = ('url', 'uuid', 'scope', 'total', 'details', 'month', 'year',
-                  'is_manually_input', 'scope_name', 'scope_type')
+        fields = ('url', 'uuid', 'scope', 'total', 'consumed', 'month', 'year',
+                  'is_manually_input', 'scope_name', 'scope_type', 'resource_type')
         read_only_fields = ('is_manually_input',)
         extra_kwargs = {
             'url': {'lookup_field': 'uuid'},
@@ -38,10 +39,15 @@ class PriceEstimateSerializer(AugmentedSerializerMixin, serializers.HyperlinkedM
         return price_estimate
 
     def get_scope_name(self, obj):
-        return six.text_type(obj.scope)  # respect to unicode
+        return six.text_type(obj.scope or obj.details.get('scope_name'))  # respect to unicode
 
     def get_scope_type(self, obj):
-        return ScopeTypeFilterBackend.get_scope_type(obj)
+        return ScopeTypeFilterBackend.get_scope_type(obj) or obj.details.get('scope_type')
+
+    def get_resource_type(self, obj):
+        if not obj.is_leaf:
+            return None
+        return SupportedServices.get_name_for_model(obj.content_type.model_class())
 
 
 class YearMonthField(serializers.CharField):