+++ nodeconductor/cost_tracking/tests/unittests/test_tasks.py
@@ -51,7 +51,7 @@ class UpdateProjectedEstimateTest(TransactionTestCase):
         update_projected_estimate(customer_uuid=self.customer.uuid.hex)
 
         reread_estimate = PriceEstimate.objects.get(id=estimate.id)
-        self.assertEqual(reread_estimate.total, self.INSTANCE_MONTHLY_COST * 2 + customer_total)
+        self.assertEqual(reread_estimate.total, self.INSTANCE_MONTHLY_COST * 2)
 
     def test_estimate_calculation_for_current_month_if_instance_already_has_estimate(self):
         now = timezone.now()
@@ -59,7 +59,7 @@ class UpdateProjectedEstimateTest(TransactionTestCase):
         estimate = PriceEstimate.objects.create(
             scope=self.instance1, month=now.month, year=now.year, total=instance_total)
 
-        update_projected_estimate(resource_uuid=self.instance1.uuid.hex)
+        update_projected_estimate(resource_str=self.instance1.to_string())
 
         reread_estimate = PriceEstimate.objects.get(id=estimate.id)
         self.assertEqual(reread_estimate.total, self.INSTANCE_MONTHLY_COST)